maint: cleanup

+++ src/common/utils/io.cpp
@@ -55,7 +55,7 @@ bool read_file(const std::string& file, std::string* data) {
     stream.seekg(0, std::ios::beg);
 
     if (size > -1) {
-      data->resize(static_cast<uint32_t>(size));
+      data->resize(static_cast<std::size_t>(size));
       stream.read(data->data(), size);
       stream.close();
       return true;
